+++ CHANGELOG.rst
@@ -7,6 +7,16 @@ The format is based on `Keep a Changelog
 <https://keepachangelog.com/en/1.0.0/>`_, and this project adheres to `Semantic
 Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
+0.7.6
+-----
+
+Added:
+
+- Added 'switch-mode' which allows user to cleanly switch between different 
+  operational modes. I.e Autocalib, EPICS and ODIN. This closes any running processes
+  and clears the shared memory.
+
+
 0.7.5
 -----